Let me tell you a story....
Ok, so I spiked on this branch and failed to commit properly.

I have added a view, controller and style for ROWs

These are now created on reception of data from the server, and added to the main view.

Whoop.

Need to sort out the TSS to make it less ugly.